remove maldidataset class
