+++ pymaldiproc/classes.py
@@ -83,60 +83,3 @@ class ConsensusPeakList(object):
         self.consensus_mz_array = consensus_peak_list_dict['mz_array']
         self.consensus_intensity_array = consensus_peak_list_dict['intensity_array']
         self.data_processing = consensus_peak_list_dict['data_processing']
-
-
-class MALDIDataset(object):
-    def __init__(self, input_path):
-        self.spectra = []
-        #self.spectra = {}
-        self.feature_matrix = None
-        self.cos_distance_matrix = None
-        self.cos_distance_matrix_shape = None
-
-        self.import_mzml(input_path)
-        #self.spectra_ids = [spectrum.spectrum_id for spectrum in self.spectra]
-        #self.spectra_ids = [spectrum_id + '|' + spectrum_replicate
-        #                    for spectrum_id, value in self.spectra.items()
-        #                    for spectrum_replicate, value in self.spectra[spectrum_id].items()]
-
-    def import_mzml(self, input_path):
-        # find mzML files
-        if input_path.endswith('.mzML'):
-            input_files = [input_path]
-        elif not input_path.endswith('.mzML') and os.path.isdir(input_path):
-            input_files = [os.path.join(dirpath, filename) for dirpath, dirnames, filenames in os.walk(input_path)
-                           for filename in filenames if filename.endswith('.mzML')]
-
-        # read in data with pyteomics
-        spectra_count = 1
-        for mzml_filename in input_files:
-            mzml_data = list(pyt.read(mzml_filename))
-            for scan_dict in mzml_data:
-                scan_dict['pymaldiproc_spectra_count'] = str(spectra_count)
-                spectrum = MALDISpectrum(scan_dict)
-                self.spectra.append(spectrum)
-                #self.spectra[spectrum.id] = {spectrum.replicate: spectrum}
-                spectra_count += 1
-
-    def get_feature_matrix(self, missing_value_imputation=True):
-        spectra_dfs_peak_picked = []
-        spectra_dfs_preprocessed = []
-        for spectrum in self.spectra:
-            spectra_dfs_peak_picked.append(pd.DataFrame(data={'mz': spectrum.peak_picked_mz_array,
-                                                              spectrum.spectrum_id: spectrum.peak_picked_intensity_array}))
-            spectra_dfs_preprocessed.append(pd.DataFrame(data={'mz': spectrum.preprocessed_mz_array,
-                                                               spectrum.spectrum_id: spectrum.preprocessed_intensity_array}))
-        self.feature_matrix = reduce(lambda x, y: pd.merge(x, y, how='outer', on='mz'), spectra_dfs_peak_picked).sort_values(by='mz')
-        if missing_value_imputation:
-            ref_matrix = reduce(lambda x, y: pd.merge(x, y, how='outer', on='mz'), spectra_dfs_preprocessed).sort_values(by='mz')
-            for colname in self.feature_matrix.columns:
-                if colname != 'mz':
-                    tmp_df = pd.merge(self.feature_matrix[['mz', colname]],
-                                      ref_matrix[['mz', colname]],
-                                      how='left',
-                                      on='mz')
-                    #self.feature_matrix[colname].fillna(tmp_df.drop('mz', axis=1).mean(axis=1), inplace=True)
-                    self.feature_matrix[colname] = tmp_df.drop('mz', axis=1).mean(axis=1).values
-
-    def get_cos_distance_matrix(self):
-        pass
