+++ pymaldiproc/preprocessing.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pandas as pd
+from functools import reduce
 from scipy.signal import savgol_filter, find_peaks, find_peaks_cwt, peak_widths
 from pyMSpec.smoothing import sg_smooth, apodization, rebin, fast_change, median
 from BaselineRemoval import BaselineRemoval
@@ -158,3 +159,28 @@ def peak_picking(list_of_spectra, method='cwt', widths=None, snr=3):
             spectrum.data_processing['peak picking']['upper peak width'] = np.max(widths)
 
     return list_of_spectra
+
+
+def get_feature_matrix(self, missing_value_imputation=True):
+    spectra_dfs_peak_picked = []
+    spectra_dfs_preprocessed = []
+    for spectrum in self.spectra:
+        spectra_dfs_peak_picked.append(pd.DataFrame(data={'mz': spectrum.peak_picked_mz_array,
+                                                          spectrum.spectrum_id: spectrum.peak_picked_intensity_array}))
+        spectra_dfs_preprocessed.append(pd.DataFrame(data={'mz': spectrum.preprocessed_mz_array,
+                                                           spectrum.spectrum_id: spectrum.preprocessed_intensity_array}))
+    self.feature_matrix = reduce(lambda x, y: pd.merge(x, y, how='outer', on='mz'), spectra_dfs_peak_picked).sort_values(by='mz')
+    if missing_value_imputation:
+        ref_matrix = reduce(lambda x, y: pd.merge(x, y, how='outer', on='mz'), spectra_dfs_preprocessed).sort_values(by='mz')
+        for colname in self.feature_matrix.columns:
+            if colname != 'mz':
+                tmp_df = pd.merge(self.feature_matrix[['mz', colname]],
+                                  ref_matrix[['mz', colname]],
+                                  how='left',
+                                  on='mz')
+                #self.feature_matrix[colname].fillna(tmp_df.drop('mz', axis=1).mean(axis=1), inplace=True)
+                self.feature_matrix[colname] = tmp_df.drop('mz', axis=1).mean(axis=1).values
+
+
+def get_cos_distance_matrix(self):
+    pass
